Inline SageMaker test fixtures, drop conftest.py
Match the repo convention: vllm/sglang/ray/vllm-omni all keep their SageMaker
endpoint fixtures in the test file rather than a per-suite conftest.py. Shared
fixtures (image_uri, region, aws_session) still come from the root
test/conftest.py.
